# cultivation-assistant/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CalibrationPointListResponse": {
        "description": "Stable list envelope for calibration points, ordered by raw value.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CalibrationPointResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "CalibrationPointListResponse",
        "type": "object"
      },
      "CalibrationPointResponse": {
        "description": "Persisted calibration point representation.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "raw_value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Raw Value",
            "type": "string"
          },
          "reservoir_id": {
            "title": "Reservoir Id",
            "type": "string"
          },
          "volume_liters": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Volume Liters",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reservoir_id",
          "raw_value",
          "volume_liters"
        ],
        "title": "CalibrationPointResponse",
        "type": "object"
      },
      "CalibrationPointsReplace": {
        "description": "Replace a reservoir's entire calibration table.",
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/CalibrationPointInput"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "points"
        ],
        "title": "CalibrationPointsReplace",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReservoirCreate": {
        "description": "Create one reservoir with its geometry.",
        "properties": {
          "capacity_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              }
            ],
            "title": "Capacity Liters"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeometryInput"
          },
          "minimum_safe_volume_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Safe Volume Liters"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "overflow_threshold_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overflow Threshold Liters"
          },
          "primary_grow_space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Grow Space Id"
          },
          "refill_threshold_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refill Threshold Liters"
          },
          "reservoir_type": {
            "$ref": "#/components/schemas/ReservoirType"
          },
          "usable_capacity_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable Capacity Liters"
          }
        },
        "required": [
          "name",
          "reservoir_type",
          "capacity_liters",
          "geometry"
        ],
        "title": "ReservoirCreate",
        "type": "object"
      },
      "ReservoirListResponse": {
        "description": "Stable list envelope for reservoirs.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReservoirSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ReservoirListResponse",
        "type": "object"
      },
...
      "ReservoirResponse": {
        "description": "Detailed reservoir representation.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "capacity_liters": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Capacity Liters",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeometryResponse"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "live_readings": {
            "items": {
              "$ref": "#/components/schemas/LiveReading"
            },
            "title": "Live Readings",
            "type": "array"
          },
          "mapping_count": {
            "default": 0,
            "title": "Mapping Count",
            "type": "integer"
          },
          "mappings": {
            "items": {
              "$ref": "#/components/schemas/ReservoirMappingResponse"
            },
            "title": "Mappings",
            "type": "array"
          },
          "minimum_safe_volume_liters": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Safe Volume Liters"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "overflow_threshold_liters": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overflow Threshold Liters"
          },
          "primary_grow_space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Grow Space Id"
          },
          "refill_threshold_liters": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refill Threshold Liters"
          },
          "reservoir_type": {
            "$ref": "#/components/schemas/ReservoirType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "usable_capacity_liters": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable Capacity Liters"
          }
        },
        "required": [
          "id",
          "name",
          "reservoir_type",
          "primary_grow_space_id",
          "capacity_liters",
          "usable_capacity_liters",
          "active",
          "geometry",
          "created_at",
          "updated_at",
          "minimum_safe_volume_liters",
          "refill_threshold_liters",
          "overflow_threshold_liters"
        ],
        "title": "ReservoirResponse",
        "type": "object"
      },
      "ReservoirSummary": {
        "description": "Compact reservoir list representation.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "capacity_liters": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Capacity Liters",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeometryResponse"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "live_readings": {
            "items": {
              "$ref": "#/components/schemas/LiveReading"
            },
            "title": "Live Readings",
            "type": "array"
          },
          "mapping_count": {
            "default": 0,
            "title": "Mapping Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "primary_grow_space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Grow Space Id"
          },
          "reservoir_type": {
            "$ref": "#/components/schemas/ReservoirType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "usable_capacity_liters": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable Capacity Liters"
          }
        },
        "required": [
          "id",
          "name",
          "reservoir_type",
          "primary_grow_space_id",
          "capacity_liters",
          "usable_capacity_liters",
          "active",
          "geometry",
          "created_at",
          "updated_at"
        ],
        "title": "ReservoirSummary",
        "type": "object"
      },
      "ReservoirType": {
        "description": "Supported physical reservoir categories.",
        "enum": [
          "autopot_reservoir",
          "dwc_bucket",
          "rdwc_control_reservoir",
          "irrigation_supply_tank",
          "mixing_tank",
          "top_off_tank",
          "ro_source_water_tank",
          "runoff_waste_tank",
          "custom_reservoir"
        ],
        "title": "ReservoirType",
        "type": "string"
      },
      "ReservoirUpdate": {
        "description": "Patch mutable reservoir fields.",
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "capacity_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Liters"
          },
          "geometry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeometryInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "minimum_safe_volume_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Safe Volume Liters"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "overflow_threshold_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overflow Threshold Liters"
          },
          "primary_grow_space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Grow Space Id"
          },
          "refill_threshold_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refill Threshold Liters"
          },
          "reservoir_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReservoirType"
              },
              {
                "type": "null"
              }
            ]
          },
          "usable_capacity_liters": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,13}0*$)\\d{0,8}\\.\\d{0,4}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable Capacity Liters"
          }
        },
        "title": "ReservoirUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "cultivation_assistant__reservoirs__schemas__EntityCandidate": {
        "description": "One role-compatible Home Assistant entity suggestion.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "compatibility": {
            "$ref": "#/components/schemas/cultivation_assistant__reservoirs__units__Compatibility"
          },
          "current_state": {
            "title": "Current State",
            "type": "string"
          },
          "device_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Class"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "friendly_name": {
            "title": "Friendly Name",
            "type": "string"
          },
          "last_updated": {
            "format": "date-time",
            "title": "Last Updated",
            "type": "string"
          },
          "source_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Unit"
          }
        },
        "required": [
          "entity_id",
          "friendly_name",
          "domain",
          "device_class",
          "source_unit",
          "current_state",
          "last_updated",
          "available",
          "compatibility",
          "explanation"
        ],
        "title": "EntityCandidate",
        "type": "object"
      },
      "cultivation_assistant__reservoirs__schemas__EntityDiscoveryResponse": {
        "description": "Stable envelope for role-filtered reservoir entity suggestions.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/cultivation_assistant__reservoirs__schemas__EntityCandidate"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "EntityDiscoveryResponse",
        "type": "object"
      },
      "cultivation_assistant__reservoirs__units__Compatibility": {
        "description": "How safely an entity unit can satisfy a reservoir role.",
        "enum": [
          "compatible",
          "convertible",
          "unknown",
          "incompatible"
        ],
        "title": "Compatibility",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Cultivation Assistant",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/home-assistant/reservoir-entities": {
      "get": {
        "operationId": "discover_reservoir_entities_api_v1_home_assistant_reservoir_entities_get",
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "required": true,
            "schema": {
              "title": "Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cultivation_assistant__reservoirs__schemas__EntityDiscoveryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover Reservoir Entities",
        "tags": [
          "reservoirs",
          "home assistant"
        ]
      }
    },
...
    "/api/v1/reservoirs": {
      "get": {
        "operationId": "list_reservoirs_api_v1_reservoirs_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservoirListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reservoirs",
        "tags": [
          "reservoirs"
        ]
      },
      "post": {
        "operationId": "create_reservoir_api_v1_reservoirs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservoirCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservoirResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Reservoir",
        "tags": [
          "reservoirs"
        ]
      }
    },
    "/api/v1/reservoirs/{reservoir_id}": {
      "delete": {
        "operationId": "archive_reservoir_api_v1_reservoirs__reservoir_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reservoir_id",
            "required": true,
            "schema": {
              "title": "Reservoir Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Reservoir",
        "tags": [
          "reservoirs"
        ]
      },
      "get": {
        "operationId": "get_reservoir_api_v1_reservoirs__reservoir_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "reservoir_id",
            "required": true,
            "schema": {
              "title": "Reservoir Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservoirResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reservoir",
        "tags": [
          "reservoirs"
        ]
      },
      "patch": {
        "operationId": "update_reservoir_api_v1_reservoirs__reservoir_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "reservoir_id",
            "required": true,
            "schema": {
              "title": "Reservoir Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservoirUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservoirResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Reservoir",
        "tags": [
          "reservoirs"
        ]
      }
    },
    "/api/v1/reservoirs/{reservoir_id}/calibration-points": {
      "get": {
        "operationId": "list_calibration_points_api_v1_reservoirs__reservoir_id__calibration_points_get",
        "parameters": [
          {
            "in": "path",
            "name": "reservoir_id",
            "required": true,
            "schema": {
              "title": "Reservoir Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationPointListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Calibration Points",
        "tags": [
          "reservoirs"
        ]
      },
      "put": {
        "operationId": "replace_calibration_points_api_v1_reservoirs__reservoir_id__calibration_points_put",
        "parameters": [
          {
            "in": "path",
            "name": "reservoir_id",
            "required": true,
            "schema": {
              "title": "Reservoir Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationPointsReplace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationPointListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Calibration Points",
        "tags": [
          "reservoirs"
        ]
      }
    },
...
  }
}